chore: move loadEnvironment to settings.ts
